Changed Salary from decimal to Integer as per the question

+++ Employee.Domain/Employee.cs
@@ -10,10 +10,10 @@ namespace Employee.Domain
 		public string Id { get; set; }
 		public string ManagerId { get; set; }
 
-		public decimal Salary { get; set; }
+		public int Salary { get; set; }
 		private Employee() { }
 
-		private Employee(string id, string managerId, decimal salary)
+		private Employee(string id, string managerId, int salary)
 		{
 			if (string.IsNullOrWhiteSpace(id)) throw new ArgumentNullException(nameof(id));
 			if(salary < 0) throw new ArgumentOutOfRangeException(nameof(salary));
@@ -21,7 +21,7 @@ namespace Employee.Domain
 			ManagerId = managerId;
 			Salary = salary;
 		}
-		public static Employee Create(string id, string managerId, decimal salary)
+		public static Employee Create(string id, string managerId, int salary)
 		{
 			return new Employee(id, managerId, salary);
 		}
